Refactor: Separate repository and service layers for Mongo and Qdrant
