+++ app/services/mongo_service.py
@@ -1,27 +1,22 @@
-from app.clients.mongo_client import MongoDBClient
+from app.repositories.mongo_repository import MongoRepository
 
 class MongoService:
-    def __init__(self, mongo_client: MongoDBClient):
-        self.db = mongo_client.get_db()
-        self.users_collection = self.db["users"]
-        self.documents_collection = self.db["documents"]
-        self.files_collection = self.db["uploaded_files"]
+    def __init__(self, mongo_repository: MongoRepository):
+        self.mongo_repository = mongo_repository
 
-    async def insert_chunks(self, chunks: list, user_id: str):
+    async def insert_chunks_with_user(self, chunks: list, user_id: str):
         for chunk in chunks:
             chunk["user_id"] = user_id
-        result = await self.documents_collection.insert_many(chunks)
-        return result.inserted_ids
+        return await self.mongo_repository.insert_chunks(chunks)
 
     async def get_user_by_username(self, username: str):
-        return await self.users_collection.find_one({"username": username})
+        return await self.mongo_repository.get_user_by_username(username)
 
     async def get_user_by_email(self, email: str):
-        return await self.users_collection.find_one({"email": email})
+        return await self.mongo_repository.get_user_by_email(email)
 
     async def create_user(self, user_data: dict):
-        return await self.users_collection.insert_one(user_data)
+        return await self.mongo_repository.create_user(user_data)
 
     async def save_file(self, file_info: dict):
-        result = await self.files_collection.insert_one(file_info)
-        return result.inserted_id
+        return await self.mongo_repository.save_file(file_info)
